+++ BCI_Processing/.gitignore
@@ -1,3 +1,5 @@
 __pycache__/*
 *.pyc
 *.txt
+
+.vscode
