+++ BCI_Processing/data/output_data/sub-P001/ses-S001/eeg/sub-P001_ses-S001_task-Default_run-001_eeg_20241223-164122_OutputFiles/PythonResponse - BCI_allExp_Report.csv
@@ -1,0 +1,14 @@
+Name:,PythonResponse
+Hostname:,Dani-Vivobook
+Creation:,1043312.511652800
+Data Type:,string
+Channel Count:,1
+Sample Count:,1105
+Nominal Sampling Rate:,0.000000000000000
+Effective Sampling Rate:,1.349869096364961
+First Adjusted Timestamp:,1043432.7479490617
+Last Adjusted Timestamp:,1044251.3458436863
+Adjusted Timestamp Space:,818.5978946245741
+First Non-Adjusted Timestamp:,1043432.7479659
+Last Non-Adjusted Timestamp:,1044251.3458603
+Non-Adjusted Timestamp Space:,818.597894400009
