+++ BCI_Processing/data/p300_example_20241223-153917_OutputFiles/PythonResponse - BCI_allExp_Report.csv
@@ -1,0 +1,14 @@
+Name:,PythonResponse
+Hostname:,DESKTOP-0CDR8QE
+Creation:,101728.2002548000
+Data Type:,string
+Channel Count:,1
+Sample Count:,3501
+Nominal Sampling Rate:,0.000000000000000
+Effective Sampling Rate:,6.308689997802698
+First Adjusted Timestamp:,101800.75359649894
+Last Adjusted Timestamp:,102355.70240497161
+Adjusted Timestamp Space:,554.9488084726763
+First Non-Adjusted Timestamp:,101800.7536018
+Last Non-Adjusted Timestamp:,102355.7024107
+Non-Adjusted Timestamp Space:,554.9488089000079
