+++ BCI_Processing/data/p300_example_20241223-153917_OutputFiles/UnityMarkerStream - LSL_Marker_Strings_allExp_Report.csv
@@ -1,0 +1,14 @@
+Name:,UnityMarkerStream
+Hostname:,DESKTOP-0CDR8QE
+Creation:,101716.7916783000
+Data Type:,string
+Channel Count:,1
+Sample Count:,1802
+Nominal Sampling Rate:,0.000000000000000
+Effective Sampling Rate:,3.3868857540589987
+First Adjusted Timestamp:,101808.89845785622
+Last Adjusted Timestamp:,102340.95065365224
+Adjusted Timestamp Space:,532.0521957960154
+First Non-Adjusted Timestamp:,101808.898463
+Last Non-Adjusted Timestamp:,102340.9506581
+Non-Adjusted Timestamp Space:,532.0521950999973
