+++ BCI_Processing/data/p300_example_20241223-153917_OutputFiles/g.USBamp-1 - EEG_allExp_Report.csv
@@ -1,0 +1,14 @@
+Name:,g.USBamp-1
+Hostname:,DESKTOP-0CDR8QE
+Creation:,101694.9595981
+Data Type:,float32
+Channel Count:,8
+Sample Count:,142207
+Nominal Sampling Rate:,256
+Effective Sampling Rate:,255.9935304150372
+First Adjusted Timestamp:,101800.65407992601
+Last Adjusted Timestamp:,102356.16031542206
+Adjusted Timestamp Space:,555.5062354960537
+First Non-Adjusted Timestamp:,101800.65405115
+Last Non-Adjusted Timestamp:,102356.16418365
+Non-Adjusted Timestamp Space:,555.5101324999996
